% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance from the fire station (in unspecified units).}
  \item{DAMAGE}{Amount of fire damage (in unspecified units).}
}
}
\source{
Original FIREDAM.csv file provided in MATH 4753 Lab 7.
}
\usage{
fire
}
\description{
A dataset containing measurements of distance from the fire station and
resulting fire damage costs. These data come from the FIREDAM.csv file used
in MATH 4753 Laboratory 7.
}
\keyword{datasets}
