% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval Generation}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations used to construct bootstrap interval}

\item{x}{A vector containing a random sample of quantitative elements}

\item{fun}{Function used to evaluate the desired population parameter}

\item{alpha}{Likelihood that true population parameter lies outside the confidence interval}

\item{cx}{Scaling parameter for text of confidence interval on the histogram}

\item{...}{Other information to add to the histogram title}
}
\value{
A histogram of the population parameter along with the desired confidence interval, as well as a list containing the confidence interval, the function used to calculate the population parameter, and the sample analyzed.
}
\description{
Bootstrap Confidence Interval Generation
}
\examples{
myboot2(iter=10000,x=c(1:10),fun="mean",alpha=0.05)
}
