% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysd.R
\name{mysd}
\alias{mysd}
\title{My Standard Deviation Function}
\usage{
mysd(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical indicating whether to remove NA values before computation.}
}
\value{
Standard deviation of the numeric vector
}
\description{
My Standard Deviation Function
}
\examples{
mysd(c(1, 2, 3, 4, 5))
}
